+++ src/components/forms/index.tsx
@@ -1,2 +1,2 @@
 export { OptInForm } from "./opt-in-form";
-export { SubscribeForm } from "./subscribe-form";
+export { ContactForm } from "./subscribe-form";
